+++ model_binding_builder/CHANGELOG.md
@@ -1,3 +1,2 @@
 ## 1.0.0
-
-- Initial version.
+- Support dataBinding and widgetBinding.